backend/requirements.txt
fastapi
psycopg2-binary
pydantic
python-decouple
SQLAlchemy
uvicorn